requirements.txt
asgiref==3.11.1
channels==4.3.0
daphne==4.2.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-filter==25.1
gunicorn==25.1.0
lxml==6.0.2
openpyxl==3.1.5
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
